Fix import fragments and complete database refactoring integration
Complete the final cleanup of import issues from automated model migration:

## Import Fragment Fixes:
- Fix "nBackend" → "CourseExecutionBackend" in course_execution_backend.py and courses.py
- Fix "nization" → "Organization" in organizations.py
- Add missing model imports in filesystem.py, permissions.py, signup.py, tests.py
- Update course_families.py with missing CourseFamily import

## Test Updates:
- Fix import paths in test_api_endpoints.py and test_model_integration.py
- Update test files to use proper ctutor_backend module prefixes
- Ensure all API and model imports work correctly

## Integration Completion:
- All 50+ files now use new SQLAlchemy model structure
- All import fragments from automated replacement resolved
- API endpoint tests passing (3/3)
- Model integration tests passing (4/4)

This completes the database refactoring integration work with all imports
properly resolved and tests passing.
