+++ src/ctutor_backend/api/filesystem.py
@@ -9,7 +9,8 @@ from ctutor_backend.interface.courses import CourseInterface, CourseProperties
 from ctutor_backend.interface.organizations import OrganizationProperties
 from ctutor_backend.interface.base import EntityInterface
 from ctutor_backend.interface.tokens import decrypt_api_key
-from ctutor_backend.model.sqlalchemy_models.course import CourseContent, CourseFamily, Organization
+from ctutor_backend.model.sqlalchemy_models.course import CourseContent, CourseFamily
+from ctutor_backend.model.sqlalchemy_models.organization import Organization
 from ctutor_backend.generator.git_helper import clone_or_pull_and_checkout
 from ctutor_backend.settings import settings
